Refactor test and test arborescence
