+++ test/assets/NM2OutputWithEstimation.js
@@ -1,0 +1,28 @@
+module.exports = {
+    '2017-04': 1200,  // prolongation
+    '2017-03': 1200,
+    '2017-02': 1200,
+    '2017-01': 1200,
+    '2016-12': 400,
+    '2016-11': 400,
+    '2016-10': 400,
+    '2016-09': 400,
+    '2016-08': 400,
+    '2016-07': 400,
+    '2016-06': 400,
+    '2016-05': 400,
+    '2016-04': 400,
+    // Les revenus estimés pour N-2 sont étalés sur 12 mois
+    '2015-12': 600,
+    '2015-11': 600,
+    '2015-10': 600,
+    '2015-09': 600,
+    '2015-08': 600,
+    '2015-07': 600,
+    '2015-06': 600,
+    '2015-05': 600,
+    '2015-04': 600,
+    '2015-03': 600,
+    '2015-02': 600,
+    '2015-01': 600,
+}
