+++ test/assets/NM2OutputWithoutEstimation.js
@@ -1,0 +1,15 @@
+module.exports = {
+    '2017-04': 1200,  // prolongation
+    '2017-03': 1200,
+    '2017-02': 1200,
+    '2017-01': 1200,
+    '2016-12': 400,
+    '2016-11': 400,
+    '2016-10': 400,
+    '2016-09': 400,
+    '2016-08': 400,
+    '2016-07': 400,
+    '2016-06': 400,
+    '2016-05': 400,
+    '2016-04': 400,
+}
